requirements.txt
selenium
pytest
pytest-html
allure-pytest
requests
